ncurses scripts adds "character" fields to tokens
Three panels are open: the leftmost displays a list of non-character NNP
Tokens. Pressing 'a' adds the topmost token to the "character build"
list displayed in the middle panel. Pressing 'c' will create a character
from the tokens in the "character build" list. The user must choose a
name and the tokens in the "character build" list will be updated with a
"character" field that has that name as a value.